api/requirements.txt
fastapi
uvicorn[standard]
psycopg[binary]

pydantic

langgraph-checkpoint-postgres
langchain
langchain-openai
langchain-mcp-adapters

python-dotenv
